Rename NodeCreateIdentity to NodeDoCreateIdentity
For consistency
