+++ pkg/latte/node_do_create_identity.go
@@ -1,0 +1,45 @@
+package latte
+
+import (
+	"context"
+
+	"github.com/authgear/authgear-server/pkg/lib/authn/identity"
+	"github.com/authgear/authgear-server/pkg/lib/workflow"
+)
+
+func init() {
+	workflow.RegisterNode(&NodeDoCreateIdentity{})
+}
+
+type NodeDoCreateIdentity struct {
+	IdentityInfo *identity.Info `json:"identity_info,omitempty"`
+}
+
+func (n *NodeDoCreateIdentity) Kind() string {
+	return "latte.NodeDoCreateIdentity"
+}
+
+func (n *NodeDoCreateIdentity) GetEffects(ctx context.Context, deps *workflow.Dependencies, w *workflow.Workflow) (effs []workflow.Effect, err error) {
+	return []workflow.Effect{
+		workflow.RunEffect(func(ctx context.Context, deps *workflow.Dependencies) error {
+			err := deps.Identities.Create(n.IdentityInfo)
+			if err != nil {
+				return err
+			}
+
+			return nil
+		}),
+	}, nil
+}
+
+func (*NodeDoCreateIdentity) CanReactTo(ctx context.Context, deps *workflow.Dependencies, w *workflow.Workflow) ([]workflow.Input, error) {
+	return nil, workflow.ErrEOF
+}
+
+func (*NodeDoCreateIdentity) ReactTo(ctx context.Context, deps *workflow.Dependencies, w *workflow.Workflow, input workflow.Input) (*workflow.Node, error) {
+	return nil, workflow.ErrIncompatibleInput
+}
+
+func (n *NodeDoCreateIdentity) OutputData(ctx context.Context, deps *workflow.Dependencies, w *workflow.Workflow) (interface{}, error) {
+	return nil, nil
+}
